refactor create temp-products controller and model
