+++ models/temp-products.model.ts
@@ -1,0 +1,111 @@
+import { db } from '../config/db';
+
+interface TempSalesItem {
+  id: number;
+  productId: number;
+  quantity: number;
+  unitPrice: number;
+  description: string;
+}
+
+interface Item {
+  productId: number; 
+  quantity: number; 
+  unitPrice: number; 
+}
+
+const getAllTemporaryProducts = async () => {
+  try {
+    const [rows] = await db.execute(
+      'SELECT t.id, p.description, t.quantity, t.unit_price FROM temporary_sales_items t JOIN products p ON t.product_id = p.id'
+    );
+    return rows as TempSalesItem[];
+  } catch (error) {
+    throw new Error('Error al consultar los productos temporales');
+  }
+};
+
+
+const insertTemporaryProduct = async ({
+  productId, 
+  quantity,
+  unitPrice 
+}: Item) => {
+  try {
+    const [result] = await db.execute(
+      `INSERT INTO temporary_sales_items (product_id, quantity, unit_price) VALUES (?, ?, ?)`,
+      [productId, quantity, unitPrice]
+    );
+    return result;
+  } catch (error) {
+    throw new Error('Error al insertar producto temporal');
+  }
+};
+
+const getTempItemByCode = async (id: number) => {
+    try {
+        const [rows] = await db.execute(
+            `SELECT p.description, t.unit_price, t.quantity
+            FROM temporary_sales_items t
+            JOIN products p ON t.product_id = p.id
+            WHERE t.id = ?`,
+            [id]
+        );
+        const items = rows as TempSalesItem[];
+        return items.length ? items[0] : null;
+    } catch (error) {
+        throw new Error('Error al consultar el producto');
+    }
+};
+
+const updateTemporaryQuantity = async (id: number, quantity: number) => {
+  try {
+// Actualiza el item en la tabla temporary_sales_items
+    await db.execute(
+      "UPDATE temporary_sales_items SET quantity = ? WHERE id = ?",
+      [quantity, id]
+    );
+
+    // Retornamos el item actualizado
+    const [rows] = await db.execute(
+      "SELECT p.description, t.unit_price, t.quantity FROM temporary_sales_items t JOIN products p ON t.product_id = p.id WHERE t.id = ?",
+      [id]
+    );
+    const items = rows as TempSalesItem[];
+    return items.length ? items[0] : null;
+  } catch (error) {
+  throw new Error('Error al editar el producto');
+  }
+};
+
+const deleteTempItemByCode = async (id: number) => {
+  try {
+  const [rowsBefore] = await db.execute(
+    'SELECT * FROM temporary_sales_items WHERE id = ?',
+    [id]
+  );
+
+const itemsBefore = rowsBefore as TempSalesItem[];
+  if (!itemsBefore.length) {
+    // No existe, devolvemos null o lanzamos error
+    return false;
+  }
+
+  // Ahora sí hacemos el DELETE
+  await db.execute(
+    'DELETE FROM temporary_sales_items WHERE id = ?',
+    [id]
+  );
+    return true;
+  } catch (error) {
+    throw new Error('Error al eliminar el producto');
+    }
+};
+
+export { 
+  getAllTemporaryProducts, 
+  insertTemporaryProduct, 
+  updateTemporaryQuantity,
+  getTempItemByCode,
+  deleteTempItemByCode
+};
